Enables std::istream to read from IOBuf chain
StreamIOBuf inherits std::basic_streambuf<char> and takes an const IOBuf pointer. StreamIOBuf can be passed in to the constructor of a standard C++ istream and the stream will source from the underlying IOBuf chain. Warning: reads on istream are all pass-by-value, so the data of the IOBuf will be copied (bummer!)